lib traceevent: Convert makefiles for stand alone libtraceevent
I had originally just copied the the make files and header files over from
the Linux kernel tools directory, but then Zamir SUN mentioned keeping the
history of these documents. I then deleted the files and recreated them
keeping their history in tact.

The paths had to be changed for the new locations that the files are now in.

Link: https://lore.kernel.org/r/7897620c-3039-5e4f-09d9-41606673fa63@gmail.com

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>