libtraceevent: Add libtraceevent.so.$(EP_VERSION) to targets and fix links
"make clean" left a broken symbolic link of libtraceevent.so.1 around,
because the clean target removes $(TARGET) and *.so, where $(TARGET) removes
the so that libtraceevent.so.1 points to, and *.so removes the symbolic link
that points to libtraceevent.so.1. But nothing removes libtraceevent.so.1.

Add both libtraceevent.so and libtraceevent.so.<major_version> to the
LIB_TARGETS variable (for cleaning them up), and also to fix the links, as
before it was:

 libtraceevent.so -> libtraceevent.so.1.1.0
 libtraceveent.so.1 -> libtraceveent.so.1.1.0

Where it should be a cascade of:

 libtraceevent.so -> libtraceevent.so.1
 libtraceevent.so.1 -> libtraceevent.so.1.1.0

Link: https://lore.kernel.org/linux-trace-devel/20201123184939.887690370@goodmis.org

Acked-by: "Tzvetomir Stoyanov (VMware)" <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>