libtraceevent: Add -ldl to the libtraceevent.so
libtraceevent.so does some dlopen() and other dynamic linking features that
requires the libdl library. There's no reason that users of libtraceevent.so
need to include -ldl if the application itself is not doing dynamic linking.

Have the build include -ldl into libtraceevent.so, such that it's in the
binary for the user to use.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>