libtraceevent: Move plugin_dir define logic to top level Makefile
The installation location of the plugin directory needs to be passed into
the C files via the -DPLUGIN_DIR=".." compiler option. But the logic for
that was only in the plugin directory, such that the main library did not
know where to find the plugins (as the PLUGIN_DIR macro was not set). This
caused the library not to load the plugins for the application.

By moving the logic to the top level Makefile and exporting the variables,
this fixes the issue.

Link: https://lore.kernel.org/linux-trace-devel/20210128163135.27ae85d5@gandalf.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>