libtraceevent: Add tep_warning()
The traceevent library uses a function with generic name warning() to
print various warning messages. This function is implemented as a weak
and can be overridden by the application. However, as the name is
generic, there could be collisions with other libraries or application,
using the functin with the same name.
Renamed warning() to tep_warning().
Renamed__vwarning to tep_vwarning().
Both functions are weak and can be overrriden. Overriding tep_warning()
affects only the warnings of libtraceevent library. The tep_vwarning()
is used by all tracing libraries: libtraceevent, libtrasefs and
libtracecmd. Overriding tep_vwarning() affects the warnings of all those
libraries.
Removed the existing __warning() function, as it is not used by the
library.

Link: https://lore.kernel.org/linux-trace-devel/20210408042623.3112002-1-tz.stoyanov@gmail.com

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>