libtraceevent: Add the API tep_parse_saved_cmdlines()
Currently, trace-cmd has a tracecmd_parse_cmdlines() function that
reads a string that holds the content of saved_cmdlines from the tracefs
file system, and then passes the pid to comm mapping to the tep handler.

There's no reason for this to be specific to trace-cmd as other
applications can easily read the saved_cmdlines with the tracefs API and
then pass it to the tep handler.

Add the same functionality to libtraceevent as tep_parse_saved_cmdlines().

Link: https://lore.kernel.org/linux-trace-devel/20210408214200.2063946-5-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>