libtraceevent: Add the API tep_parse_printk_formats()
Currently, trace-cmd has a tracecmd_parse_ftrace_printk() function that
reads a string that holds the content of printk_formats from the tracefs
file system, and then passes the address of the printk strings mappings
to the tep handler.

There's no reason for this to be specific to trace-cmd as other
applications can easily read the printk_formats with the tracefs API and
then pass it to the tep handler.

Add the same functionality to libtraceevent as tep_parse_printk_formats().

Link: https://lore.kernel.org/linux-trace-devel/20210408214200.2063946-6-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>