libtraceevent: Add reading func_repeats events to function plugin
With the new ftrace event func_repeats coming to the kernel, add processing
the output of it so it fits right in with the function tracer.

Since it uses the same event fields as the function trace, it can call the
function handler first, which will process the func_repeats event just like
a function, and then print the count and convert the delta to display the
actual timestamp of the last function that repeated.

Add a weak variable called "tep_func_repeat_format" to allow applications
to override how the timestamp of the last function is displayed.

Link: https://lore.kernel.org/linux-trace-devel/20210413170756.3c4638f9@gandalf.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>