libtraceevent: Add new line to trace-seq warnings
The WARN_ONCE() logic in trace_seq is based off of the WARN_ONCE() logic
in the Linux kernel, where the message does not require to end with a new
line. But the implementation is different than the kernel, and it lacks
adding a new line when it writes the output.

Add the new line to make it match what happens in the Linux kernel.

Link: https://lore.kernel.org/linux-trace-devel/20210514122825.162203-1-y.karadz@gmail.com/
Link: https://lore.kernel.org/linux-trace-devel/20210514093858.4e862bda@gandalf.local.home

Fixes: 3b760f95 ("tools include: Move perf's bug.h to a generic place")
Reported-by: "Yordan Karadzhov (VMware)" <y.karadz@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>