libtraceevent: Have the header protection be more name space safe
The header protection for event-utils.h is just _UTIL_H which is not very
name space safe and if another header were to use that, it could cause
this header not to be included (or this header could cause the other
header not to be included). Also update event-parse.h as well.

Link: https://lore.kernel.org/linux-trace-devel/20210514205036.1383167-3-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>