libtraceevent: Free "->" token in process_entry()
Fixing an issue with parsing "(REC)->" when it expected "REC->", changed
the test of "->" to read into a "token" variable instead of just
"expecting" it. This was needed to test for the ")" case. But since it is
now read into a token, that token needs to be freed. But that was missed.

Free the token.

This was found by the libtracefs "make test_mem" tests.

Link: https://lore.kernel.org/linux-trace-devel/20210623224730.6af93f6c@rorschach.local.home

Fixes: 62823da1 ("libtraceevent: Handle parsing of "(REC)->" case")
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>