libtraceevent: Do not print message if wakeup success field is missing
The "success" field for sched wakeup has been hardcoded as "1" for years
now, because it is no longer anything but successful. The sched wakeup
plugin had the "err" parameter of tep_get_field_val() backwards, where it
had no message for fields it expects to see, but will show a error message
for the success field, which may not be there.

Link: https://lore.kernel.org/linux-trace-devel/20210820173559.5644c2b0@oasis.local.home

Fixes: 4f6857aa ("tools lib traceevent: Add sched_switch plugin")
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>