libtraceevent: A new API for trace page size
Currently, the size of a trace buffer page is equal to the size of the
system memory page. This may change in the future, so this assumption
may not be valid. The proper way to determine the trace buffer page size
is by using the information from "events/header_page" ftrace file:
 ...
 field: char data;	offset:16;	size:4080;	signed:1;
 ...
The trace buffer page size is the size of the "data" filed + its offset.

A new libtraceevent API is introduced, for getting the trace buffer page
size using that formula:
 tep_get_sub_buffer_size()

Note, that some old kernels may not have "events/header_page" file. For
those, the system memory page size is returned.

Link: https://lore.kernel.org/linux-trace-devel/20211126034606.190816-1-tz.stoyanov@gmail.com

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>