libtraceevent: Fix installing of man pages in remote directory
When the building of the man pages is done in a directory other than the
repo directory (with the O= option), the intermediate build file name
incorrectly appends a "install-" prefix to the path and not to the file,
causing the installation to fail.

Fix it by using patsubst instead of addprefix to append the "install-"
name to the intermediate files for the installation.

Link: https://lore.kernel.org/linux-trace-devel/20211216213956.13934-2-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>