libtraceevent: Rewrite Makefiles to be like libtracefs
The libtraceevent code has moved around various homes. It originated in
the trace-cmd.git repository and then moved to the Linux kernel under
tools/libs. Finally it made its home in its own libtraceevent git repo.
During its time at different foster homes, it took with it various ways to
build it from the other homes and this was incorporated into the
Makefiles. But the kernel Makefiles were rather complex as it had to build
not only libtraceevent but other tooling.

Unfortunately, that complexity made it's way into libtraceevent, making it
difficult to add new functionality. Instead, rewrite it to be more like
libtracefs that has a much simpler Makefile structure.

Link: https://lore.kernel.org/linux-trace-devel/20211216213956.13934-3-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>