libtraceveent: Add testing of old dynamic string event layout
When running against old trace.dat files, the output would not print
strings. This was due to a change that did not take into account the old
dynamic format that did not have strings as an array.

Add a unit test to cover this case so that it does not break again.

Link: https://lore.kernel.org/linux-trace-devel/20211216213956.13934-7-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>