libtraceevent: Replace tep_print_field() with tep_print_field_content()
The API tep_print_field() does not let the user define the size of the
payload that is being parsed. This can lead to dangerous code if the data
itself is untrusted. It could define a string and offset to read other
parts of the application.

Add a new tep_print_field_content() that adds the size variable that can
be used to prevent accesses outside the data.

Note, there is still many ways to circumvent the size limitation that
needs to be fixed, but having an API that will not let you fix it is a
problem.

Also fixed at formatting issue with print_field() that had void before the
"static inline".

Link: https://lore.kernel.org/linux-trace-devel/20211216213956.13934-10-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>