libtraceevent: Have print_field_raw() handle old data layout
Old kernels would produce dynamic strings with only a two byte offset. The
length of the array was determined by the string length. When parsing, if
the field is dynamic with a size of 2, then it is the old dynamic string
layout. Set the field to both a string and an array.

Link: https://lore.kernel.org/linux-trace-devel/20211216213956.13934-11-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>