libtraceevent: Add warnings if fields are outside the event
If a field is referenced outside of the event that is being processed,
do not access that field and warn about it.

This was triggered when there was a bug in the kernel that showed the
wrong offset to the fields that were outside the size of the event. When
trace-cmd used this library to get a field of an event that was at the
end of a sub buffer, it read past the page that was not allocated and
crashed.

Not only warn when this happens, but also just return a zero value and
not read past the event.

Link: https://lore.kernel.org/all/20220731015928.7ab3a154@rorschach.local.home/
Link: https://lore.kernel.org/linux-trace-devel/20220731120247.71528711@rorschach.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>