libtraceevent: Fix uninitialized variable in eval_num_arg()
It appears that my compiler version did not notice that "field_size"
passed to check_data_offset_size() in the TEP_PRINT_DYNAMIC_ARRAY case was
not initialized. But other compilers do.

Since this case we just want to make sure the offset is not passed the end
of the event data, we can simply pass in zero.

Link: https://lore.kernel.org/linux-trace-devel/20220923123014.489e0e39@gandalf.local.home

Fixes: efd32896dd5db ("libtraceevent: Add warnings if fields are outside the event")
Reported-by: Ross Zwisler <zwisler@google.com>
Reviewed-by: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>