libtraceevent: Add tep_find_function_info()
I needed a way to get the size of a function, and instead of creating
another function that gets the size (having to look up the function by
address again), I decided it would be better to have a single function
that retrieved the name, starting address and the size.

This implements tep_find_function_info() that does just that.

Link: https://lore.kernel.org/linux-trace-devel/20221202180625.58262735@gandalf.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>