libtraceevent kbuffer: Add SAME_AS_HOST for endian and size
Add a SAME_AS_HOST to the endian and size parameters of kbuffer_alloc()
that makes it easier for the applications that simply read the ring
buffer.

If the host has 4 byte word, it will call uname() and check the machine
value. If machine contains the string "64", it will assume that the kernel
is 64 bit and that is what the host will be considered.

Link: https://lore.kernel.org/linux-trace-devel/20221208163644.185bf199@gandalf.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>