libtraceeevent: Add kbuffer API man pages
The kbuffer API has never been documented. Add the man pages for the
functions that applications are likely to use.

Note, the functions:

  kbuffer_set_old_format
  kbuffer_raw_get
  kbuffer_ptr_delta
  kbuffer_translate_data

Are for internal use by trace-cmd, and really should not be used by other
applications as they are implementation specific. They are exported, so
they should be stable, but lets not encourage their use by documenting how
to use them.

Link: https://lore.kernel.org/linux-trace-devel/20221208190709.1f1dc458@gandalf.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>