libtraceevent: Calculate dynamic size for dynamic fields as well as arrays
With the new "cpumask_t" field, which is dynamic and is denoted with
"__data_loc", it needs to be processed as dynamic in size, and not as
static.

When calculating the field size, check for both the field being an array
as well as being dynamic, as it can be dynamic without being an array.

Link: https://lore.kernel.org/linux-trace-devel/20221213221651.0d36897a@gandalf.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>