libtraceevent: Allow stack traces to show offset too
The stack traces only show the function name and the full address, which
can be useless for running under gdb, because the address could be
randomized, and to find the locations within the functions, the offsets
from the start of the function really need to be displayed and not just
the full address, as the address isn't known by gdb.

Have the function plugin "offset" option affect stack traces too. That
way, there's a way to display the offsets of the functions too.

Link: https://lore.kernel.org/linux-trace-devel/20221219224258.68a86565@gandalf.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>