libtraceevent: Remove tep_find_function{_address}() from libtraceevent-func_apis.txt
When a man page has more than 9 APIs described in it, asciidoc starts to
complain. As there were more functions being added to the function tep
interface, the man page that described the function interface was split to
have the "find" functions in a separate man page. But two functions were
left behind in the old libtraceevent-func_apis.txt man page:

  tep_find_function()
  tep_find_function_address()

This causes issue because it becomes ambiguous to which man page to read,
and one may become out of date. Remove the description from the
libtraceevent-func_apis.txt as they are described in
libtraceevent-func_find.txt man page.

Link: https://lore.kernel.org/linux-trace-devel/20221224153225.ojre2c3fxktfxtj2@carbon.lan/
Link: https://lore.kernel.org/linux-trace-devel/20230104172655.145516-3-rostedt@goodmis.org

Fixes: dd8094ac7252 ("tools lib traceevent: Man pages for function related libtraceevent APIs")
Reported-by: Daniel Wagner <dwagner@suse.de>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>