libtraceevent: Fix output of raw prints
The raw prints that uses the parsed fields directly, had a bug in it where
the check to catch if reading the event went beyond the event size it
would warn. But instead of testing against the event size, it was testing
against the field size. The test was suppose to test:

  field->offset + field->size > data_size

Which would catch an overflow, but instead it was testing:

  field->offset + field->size > field->size

Which will always be true! (well, if the field was not at the beginning of
the data, which is always is due to meta data).

Have it check the data size and not the field size.

Link: https://lore.kernel.org/linux-trace-devel/20230106142341.15df4486@gandalf.local.home

Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=216896
Reported-by:  Douglas RAILLARD  <douglas.raillard@arm.com>
Fixes: 09f02890358a2 ("libtraceevent: Improve tep_print_field()")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>