libtraceevent plugins: Parse sched_switch "prev_state" field for state info
The write_state() function uses a hard coded string "SDTtXZPI" to index
the sched_switch prev_state field bitmask. This is fine, except for when
the kernel changes this string, in which case this will break again.

Worse yet, there can be various saved trace files that have various
versions of this string, and updating the string may work for one trace
file, it will likely break another trace file.

Instead, look into the event itself, and how it parsed the "print fmt".
Using the tep_print_args, the mapping between the bits and the output that
the kernel uses is exposed to user space. Walk the print arguments until
the __print_flags() for the "prev_state" field is found, and use that to
build the states string for future parsing.

Save the "prev_state_field" pointer, as it should be the same for later
occurrences, but if more than one trace data (more than one tep handler)
is being parsed, the string will need to be updated each time a new field
is passed in, as this is not saved in the tep handle itself.

Link: https://lore.kernel.org/linux-trace-devel/20231224140732.7d41698d@rorschach.local.home

Cc: Ze Gao <zegao2021@gmail.com>
Cc: Ze Gao <zegao@tencent.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>