libtraceevent: Fix tep_kbuffer() to have kbuf assign long_size
If the tep handle that returns the kbuf from tep_kbuffer() did not have its
long_size initialized, neither will the kbuffer it returns. This can cause
inconsistent results. Default the long_size to the "commit" size of the header
page as that should also be the size of long.

Link: https://lore.kernel.org/linux-trace-devel/20231224191813.1076074-3-rostedt@goodmis.org

Fixes: 21ba6336 ("libtraceevent: Add kbuffer_create()")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>