kbuffer: Add kbuffer_read_buffer()
The kbuffer_read_buffer() function will fill the given buffer from the
kbuf the same way the kernel would do a read system call. That is, if the
length len is less than the sub buffer size, or the current index is
non-zero, it will start from the next event to read, and create buffer as a new
sub buffer (with a timestamp and commit header) with that event that was found
and including all events after that can fit within len.

The len must include the size of the sub buffer header as
well as the events to include. That is, len is the allocate size of buffer
that can be filled.  The return from this function is the index of the end of
the last event that was added.  If there are no more events then zero is
returned, and if the buffer can not copy any events because len was too small,
then -1 is returned.

Link: https://lore.kernel.org/linux-trace-devel/20231224191813.1076074-4-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>