kbuffer: Add kbuffer_refresh() API
Add a way to refresh the current size of the loaded subbuffer of the
kbuffer. This is needed if the loaded subbuffer has a writer on it and it
needs to update for new events that have been written.

Note, no memory barriers are used here and that would be required by the
application.

Link: https://lore.kernel.org/linux-trace-devel/20231224191813.1076074-7-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>